% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Estimation of Sample Statistics}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of samples to take from original sample}

\item{x}{Original sample to take values from}

\item{fun}{Function name to calculate a statistic ("mean", "sd", etc)}

\item{alpha}{Confidence level (0.05, 0.1, etc)}

\item{cx}{Size factor of numerical values on the graph}

\item{...}{Additional text to place on the graph}
}
\value{
A histogram containing a confidence interval and point estimate for the sample statistic
}
\description{
Bootstrap Estimation of Sample Statistics
}
\examples{
myboot2(x=rnorm(25,mean=25,sd=10),xlab="mean")
}
