% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95\% Confidence Interval for Mean}
\usage{
myci(x)
}
\arguments{
\item{x}{A sample vector from which to estimate a population mean}
}
\value{
A 95\% confidence interval for the mean, and an invisible list of key values
}
\description{
95\% Confidence Interval for Mean
}
\examples{
myci(rnorm(2,mean=5,sd=4))
}
