% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve and Probability}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Lower tail probability cutoff}
}
\value{
A normal curve shaded with a lower tail probability, and a probability value
}
\description{
Normal Curve and Probability
}
\examples{
myncurve(10,4,9)
}
