% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Flight Overbooking Seats}
\usage{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
\arguments{
\item{N}{Number of seats on a flight}

\item{gamma}{Probability of overbooking the flight ("pain" factor)}

\item{p}{Probability of a ticket holder showing}
}
\value{
2 plots and a list of values
}
\description{
Flight Overbooking Seats
}
\examples{
ntickets(N=400,gamma=0.02,p=0.95)
}
